[refactor] 패키지 관리
2개의 DialogFragment 패키지 이동
