+++ Android/FunBox/app/src/main/java/com/example/funbox/presentation/game/quiz/NetworkAlertFragment.kt
@@ -1,9 +1,9 @@
-package com.example.funbox
+package com.example.funbox.presentation.game.quiz
 
 import androidx.fragment.app.activityViewModels
+import com.example.funbox.R
 import com.example.funbox.databinding.FragmentNetworkAlertBinding
 import com.example.funbox.presentation.BaseDialogFragment
-import com.example.funbox.presentation.game.quiz.QuizViewModel
 
 class NetworkAlertFragment : BaseDialogFragment<FragmentNetworkAlertBinding>(R.layout.fragment_network_alert) {
