+++ Android/FunBox/app/src/main/java/com/example/funbox/presentation/game/quiz/ScoreBoardFragment.kt
@@ -1,9 +1,9 @@
-package com.example.funbox
+package com.example.funbox.presentation.game.quiz
 
 import androidx.fragment.app.activityViewModels
+import com.example.funbox.R
 import com.example.funbox.databinding.FragmentScoreBoardBinding
 import com.example.funbox.presentation.BaseDialogFragment
-import com.example.funbox.presentation.game.quiz.QuizViewModel
 
 
 class ScoreBoardFragment : BaseDialogFragment<FragmentScoreBoardBinding>(R.layout.fragment_score_board) {
